Show test results in test results tool window and get multiline test output only for adaptive courses Show balloons with task check result for adaptive courses

+++ python/educational-core/student/src/com/jetbrains/edu/learning/checker/StudySmartChecker.java
@@ -14,6 +14,7 @@ import com.jetbrains.edu.learning.core.EduDocumentListener;
 import com.jetbrains.edu.learning.core.EduNames;
 import com.jetbrains.edu.learning.core.EduUtils;
 import com.jetbrains.edu.learning.courseFormat.AnswerPlaceholder;
+import com.jetbrains.edu.learning.courseFormat.Course;
 import com.jetbrains.edu.learning.courseFormat.StudyStatus;
 import com.jetbrains.edu.learning.courseFormat.TaskFile;
 import com.jetbrains.edu.learning.StudyTaskManager;
@@ -67,14 +68,17 @@ public class StudySmartChecker {
         Process smartTestProcess = testRunner.createCheckProcess(project, windowCopy.getPath());
         final CapturingProcessHandler handler = new CapturingProcessHandler(smartTestProcess, null, windowCopy.getPath());
         final ProcessOutput output = handler.runProcess();
-        boolean res = StudyTestsOutputParser.getTestsOutput(output).isSuccess();
-        StudyTaskManager.getInstance(project).setStatus(userAnswerPlaceholder, res ? StudyStatus.Solved : StudyStatus.Failed);
-        StudyUtils.deleteFile(windowCopy);
-        if (fileWindows != null) {
-          StudyUtils.deleteFile(fileWindows);
-        }
-        if (!resourceFile.delete()) {
-          LOG.error("failed to delete", resourceFile.getPath());
+        final Course course = StudyTaskManager.getInstance(project).getCourse();
+        if (course != null) {
+          boolean res = StudyTestsOutputParser.getTestsOutput(output, course.isAdaptive()).isSuccess();
+          StudyTaskManager.getInstance(project).setStatus(userAnswerPlaceholder, res ? StudyStatus.Solved : StudyStatus.Failed);
+          StudyUtils.deleteFile(windowCopy);
+          if (fileWindows != null) {
+            StudyUtils.deleteFile(fileWindows);
+          }
+          if (!resourceFile.delete()) {
+            LOG.error("failed to delete", resourceFile.getPath());
+          }
         }
       }
     }
